Move ParameterUtils to presto-analyzer
