+++ presto-analyzer/src/main/java/com/facebook/presto/sql/analyzer/utils/ParameterExtractor.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.execution;
+package com.facebook.presto.sql.analyzer.utils;
 
 import com.facebook.presto.sql.tree.DefaultTraversalVisitor;
 import com.facebook.presto.sql.tree.LambdaArgumentDeclaration;
