+++ presto-main/src/main/java/com/facebook/presto/sql/rewrite/DescribeInputRewrite.java
@@ -41,7 +41,6 @@ import java.util.Map;
 import java.util.Optional;
 
 import static com.facebook.presto.common.type.UnknownType.UNKNOWN;
-import static com.facebook.presto.execution.ParameterExtractor.getParameters;
 import static com.facebook.presto.sql.QueryUtil.aliased;
 import static com.facebook.presto.sql.QueryUtil.ascending;
 import static com.facebook.presto.sql.QueryUtil.identifier;
@@ -50,6 +49,7 @@ import static com.facebook.presto.sql.QueryUtil.row;
 import static com.facebook.presto.sql.QueryUtil.selectList;
 import static com.facebook.presto.sql.QueryUtil.simpleQuery;
 import static com.facebook.presto.sql.QueryUtil.values;
+import static com.facebook.presto.sql.analyzer.utils.ParameterExtractor.getParameters;
 import static com.facebook.presto.util.AnalyzerUtil.createParsingOptions;
 import static java.util.Objects.requireNonNull;
